- Correctly Fixed Aligned Heap Allocation Bugs - Store Alignment in Header for Defragmentation use - Added Moveable Heap Header for Start work on Defragmentable Heap - Added Function to Global Time for reading the current frame time. - Added Draw Quad Functionality to IMGUIWindow class - Fixed Particle System Memory Leak - Moved Prewarming of Particle Systems to Constructor, should only happen once. - Added Heap Fragmentation Visualisation to ProfilerWindow - Fixed Bug in Vertex Processing Scenario preventing re-run.

+++ Memory/src/ScenarioManager.cpp
@@ -21,7 +21,7 @@ Scenarios({nullptr, nullptr})
 #ifdef x64
 	config.Bootup.TotalSizeToLoad = 2 * GB;
 #else
-	config.Bootup.TotalSizeToLoad = 0.9 * GB;
+	config.Bootup.TotalSizeToLoad = (size_t)0.9 * GB;
 #endif
 
 	config.Gameplay.LoadInterval = 0.5f;
